requirements.txt
pyzmq
git+https://github.com/coagulant/progressbar-python3.git